+++ web/vite.config.js
@@ -8,9 +8,6 @@ const amplipiurl = process.env.AMPLIPI_URL || "http://127.0.0.1";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  build: {
-    target: 'es2022'
-  },
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),